[security] Implement nonce reuse protection for AEAD cipher modes
Add nonce reuse protection to prevent cryptographic vulnerabilities in AEAD
cipher modes (GCM, CCM, EAX, OCB).

Security Enhancements:
- Nullify IV state after encryption in base engineReset() to force re-initialization
- Add IV validation in engineUpdate() and startAeadMode() to reject operations without proper initialization
- Throw IllegalStateException when attempting to reuse cipher without re-initialization
- Remove conditional IV check in BotanBaseSymmetricCipher to always call botan_cipher_start

AEAD Cipher Implementation Changes:
- Refactor base engineReset() to nullify IV in encryption mode only
- Add AAD buffering to defer botan_cipher_set_associated_data() until first cipher operation
- Implement proper state management with isInitialized flag
- Add custom engineReset() for AES-SIV to leverage its nonce-misuse resistance
- Ensure consistent buffer management (payload_buffer and aad_buffer)

Structural Changes:
- Move AEAD cipher classes from seckey/aead to seckey/block/aead package
- Update package references in BotanProvider (AEAD_CIPHER_PREFIX)
- Add BOTAN_ENCRYPT_MODE and BOTAN_DECRYPT_MODE constants
- Reorganize test resources to match new package structure

Test Coverage:
- Add comprehensive nonce reuse protection tests for all AEAD modes
- Verify IllegalStateException on multiple doFinal() without re-initialization
- Verify IllegalStateException on update() after doFinal() without re-initialization
- Verify decryption mode allows multiple operations (expected behavior)
- Test coverage: GCM, CCM, EAX, OCB modes
